Move the new functionality into openapi/ servlet
