--- metadata-service/operations-servlet/src/main/java/com/datahub/operations/config/SpringWebConfig.java
@@ -1,33 +1,0 @@
-package com.datahub.operations.config;
-
-import io.swagger.v3.oas.annotations.OpenAPIDefinition;
-import io.swagger.v3.oas.annotations.servers.Server;
-import java.util.List;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.format.FormatterRegistry;
-import org.springframework.http.converter.ByteArrayHttpMessageConverter;
-import org.springframework.http.converter.FormHttpMessageConverter;
-import org.springframework.http.converter.HttpMessageConverter;
-import org.springframework.http.converter.StringHttpMessageConverter;
-import org.springframework.http.converter.json.MappingJackson2HttpMessageConverter;
-import org.springframework.web.servlet.config.annotation.EnableWebMvc;
-import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
-
-
-@EnableWebMvc
-@OpenAPIDefinition(servers = {@Server(url = "/welcome/", description = "Default Server URL")})
-@Configuration
-public class SpringWebConfig implements WebMvcConfigurer {
-
-  @Override
-  public void configureMessageConverters(List<HttpMessageConverter<?>> messageConverters) {
-    messageConverters.add(new StringHttpMessageConverter());
-    messageConverters.add(new ByteArrayHttpMessageConverter());
-    messageConverters.add(new FormHttpMessageConverter());
-    messageConverters.add(new MappingJackson2HttpMessageConverter());
-  }
-
-  @Override
-  public void addFormatters(FormatterRegistry registry) {
-  }
-}
